refactor(PosixMq/Queue): Avoid use of mq_getattr
Using the mq_getattr method safely via JNA seems to be non-trivial.

This change refactors the exposed high-level methods to be more
explicit, this means that opening a queue now requires the user to
pass in the queue size.

Creating a queue will default to the system default queue size.

Inside of an application that creates and uses queues with this library
the sizes will match if arguments are left to defaults, but when
interfacing with external programs the user should ensure that the
sizes match.